# legion/operator/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  routes.HTTPResult:
    properties:
      message:
        description: Success of error message
        type: string
    type: object
...
  v1.TokenRequest:
    properties:
      expiration_date:
        description: Explicitly set expiration date for token
        type: string
      md_name:
        description: Model Deployment name
        type: string
    type: object
  v1.TokenResponse:
    properties:
      token:
        type: string
    type: object
...
host: '{{.Host}}'
info:
  contact: {}
  description: This is a EDI server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EDI API
  version: "1.0"
paths:
...
  /api/v1/model/token:
    post:
      consumes:
      - application/json
      description: Create a JWT token for access to the model service
      parameters:
      - description: Create a model JWT token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/v1.TokenRequest'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.TokenResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.HTTPResult'
            type: object
      summary: Create a model JWT token
      tags:
      - Token
...
swagger: "2.0"